CantinaProjeto/backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pyjwt==2.12.1
werkzeug==3.1.6
gunicorn==26.0.1
